+++ Assets/Scripts/Drone/Tutorial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c9b011028ea0b62469d3774240a2e87a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
