+++ Assets/Scripts/Drone/Tutorial/DroneMovementTutorial.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92aae68a1ee78ca479970923635e4033
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
